# eventos-vivos-web/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "EventosVivos.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Eventos": {
      "post": {
        "tags": [
          "Eventos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Evento"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Evento"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Evento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Eventos"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Eventos/filtrar": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TipoEvento"
            }
          },
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EstadoEvento"
            }
          },
          {
            "name": "fechaInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fechaFin",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "titulo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Eventos/{id}": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Eventos/reporte-ocupacion/{eventoId}": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservas": {
      "post": {
        "tags": [
          "Reservas"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reserva"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Reserva"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Reserva"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservas/confirmar/{id}": {
      "put": {
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservas/cancelar/{id}": {
      "put": {
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Reservas/evento/{eventoId}": {
      "get": {
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "name": "eventoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Venues": {
      "post": {
        "tags": [
          "Venues"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Venue"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Venue"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Venue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Venues"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EstadoEvento": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "EstadoReserva": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "Evento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "titulo": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "venueId": {
            "type": "integer",
            "format": "int32"
          },
          "venue": {
            "$ref": "#/components/schemas/Venue"
          },
          "capacidadMaxima": {
            "type": "integer",
            "format": "int32"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date-time"
          },
          "fechaFin": {
            "type": "string",
            "format": "date-time"
          },
          "precioEntrada": {
            "type": "number",
            "format": "double"
          },
          "tipo": {
            "$ref": "#/components/schemas/TipoEvento"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoEvento"
          },
          "reservas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reserva"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Reserva": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "eventoId": {
            "type": "integer",
            "format": "int32"
          },
          "evento": {
            "$ref": "#/components/schemas/Evento"
          },
          "usuarioId": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "nombreComprador": {
            "type": "string",
            "nullable": true
          },
          "emailComprador": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoReserva"
          },
          "codigoReserva": {
            "type": "string",
            "nullable": true
          },
          "fechaCancelacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fechaPago": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TipoEvento": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Venue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "capacidad": {
            "type": "integer",
            "format": "int32"
          },
          "ciudad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
  "tags": [
    {
      "name": "Eventos"
    },
    {
      "name": "Reservas"
    },
    {
      "name": "Venues"
    }
  ]
}